Separate download button from the variant viewer visualisation
